Conflict

+++ src/main/java/top/jwmc/kuri/ezdrawboard/client/EnhancedDrawingBoard.java
@@ -50,7 +50,7 @@ public class EnhancedDrawingBoard extends Application {
             this.y2 = y2;
             this.brushSize = brushSize;
         }
-        public Drawing(ToolType type, Color color,List<Point2D> path, double x1, double y1, double x2, double y2, int brushSize) {
+        public Drawing(ToolType type, Color color,List<Point2D> path, double x1, double y1, double x2, double y2, double brushSize) {
             this.type = type;
             this.color = color;
             this.path = path;
